Move the validation of organization social profiles in the helper
