+++ src/integrations/admin/social-profiles-helper.php
@@ -2,6 +2,8 @@
 
 namespace Yoast\WP\SEO\Integrations\Admin;
 
+use Yoast\WP\SEO\Helpers\Options_Helper;
+
 /**
  * Class Social_Profiles_Helper.
  */
@@ -26,12 +28,30 @@ class Social_Profiles_Helper {
 	];
 
 	/**
-	 * Gets the social profiles fields names.
+	 * The fields for the organization social profiles payload.
 	 *
-	 * @return string[] The social profiles fields names.
+	 * @var array
 	 */
-	public function get_person_social_profiles_fields() {
-		return $this->person_social_profile_fields;
+	private $organization_social_profile_fields = [
+		'facebook_site'     => 'get_non_valid_url',
+		'twitter_site'      => 'get_non_valid_twitter',
+		'other_social_urls' => 'get_non_valid_url_array',
+	];
+
+	/**
+	 * The Options_Helper instance.
+	 *
+	 * @var Options_Helper
+	 */
+	protected $options_helper;
+
+	/**
+	 * Social_Profiles_Helper constructor.
+	 *
+	 * @param Options_Helper $options_helper The WPSEO options helper.
+	 */
+	public function __construct( Options_Helper $options_helper ) {
+		$this->options_helper = $options_helper;
 	}
 
 	/**
@@ -59,7 +79,7 @@ class Social_Profiles_Helper {
 	}
 
 	/**
-	 * Stores the values for the social profiles.
+	 * Stores the values for the person's social profiles.
 	 *
 	 * @param int   $person_id       The id of the person to edit.
 	 * @param array $social_profiles The array of the person's social profiles.
@@ -79,6 +99,61 @@ class Social_Profiles_Helper {
 	}
 
 	/**
+	 * Stores the values for the organization's social profiles.
+	 *
+	 * @param array $social_profiles An array with the social profiles values to be saved in the db.
+	 *
+	 * @return string[] An array of field names which failed to be saved in the db.
+	 */
+	public function set_organization_social_profiles( $social_profiles ) {
+		$failures         = [];
+		$invalid_settings = [];
+
+		// First validate all social profiles, before even attempting to save them.
+		foreach ( $this->organization_social_profile_fields as $field_name => $validation_method ) {
+			if ( ! isset( $social_profiles[ $field_name ] ) ) {
+				$invalid_settings[] = $field_name;
+				continue;
+			}
+
+			$value_to_validate = $social_profiles[ $field_name ];
+			$invalid_settings  = \array_merge( $invalid_settings, \call_user_func( [ $this, $validation_method ], $value_to_validate, $field_name ) );
+		}
+
+		if ( ! empty( $invalid_settings ) ) {
+			return $invalid_settings;
+		}
+
+		// All social profiles look good, now let's try to save them.
+		foreach ( array_keys( $this->organization_social_profile_fields ) as $field_name ) {
+			if ( isset( $social_profiles[ $field_name ] ) ) {
+				$result = $this->options_helper->set( $field_name, $social_profiles[ $field_name ] );
+				if ( ! $result ) {
+					/**
+					 * The value for Twitter might have been sanitised from URL to username.
+					 * If so, $result will be false. We should check if the option value is part of the received value.
+					 */
+					if ( $field_name === 'twitter_site' ) {
+						$current_option = $this->options_helper->get( $field_name );
+						if ( ! \strpos( $social_profiles[ $field_name ], 'twitter.com/' . $current_option ) ) {
+							$failures[] = $field_name;
+						}
+					}
+					else {
+						$failures[] = $field_name;
+					}
+				}
+			}
+		}
+
+		if ( ! empty( $failures ) ) {
+			return $failures;
+		}
+
+		return [];
+	}
+
+	/**
 	 * Checks if the current user has the capability to edit a specific user.
 	 *
 	 * @param int $person_id The id of the person to edit.
@@ -87,5 +162,57 @@ class Social_Profiles_Helper {
 	 */
 	public function can_edit_profile( $person_id ) {
 		return \current_user_can( 'edit_user', $person_id );
+	}
+
+	/**
+	 * Checks if url is not valid and returns the name of the setting if it's not.
+	 *
+	 * @param string $url         The url to be validated.
+	 * @param string $url_setting The name of the setting to be updated with the url.
+	 *
+	 * @return array An array with the setting that the non-valid url is about to update.
+	 */
+	protected function get_non_valid_url( $url, $url_setting ) {
+		if ( $this->options_helper->validate_social_url( $url ) ) {
+			return [];
+		}
+
+		return [ $url_setting ];
+	}
+
+	/**
+	 * Checks if urls in an array are not valid and return the name of the setting if one of them is not, including the non-valid url's index in the array
+	 *
+	 * @param array  $urls         The urls to be validated.
+	 * @param string $urls_setting The name of the setting to be updated with the urls.
+	 *
+	 * @return array An array with the settings that the non-valid urls are about to update, suffixed with a dash-separated index of the positions of those settings, eg. other_social_urls-2.
+	 */
+	protected function get_non_valid_url_array( $urls, $urls_setting ) {
+		$non_valid_url_array = [];
+
+		foreach ( $urls as $key => $url ) {
+			if ( ! $this->options_helper->validate_social_url( $url ) ) {
+				$non_valid_url_array[] = $urls_setting . '-' . $key;
+			}
+		}
+
+		return $non_valid_url_array;
+	}
+
+	/**
+	 * Checks if the twitter value is not valid and returns the name of the setting if it's not.
+	 *
+	 * @param array  $twitter_site    The twitter value to be validated.
+	 * @param string $twitter_setting The name of the twitter setting to be updated with the value.
+	 *
+	 * @return array An array with the setting that the non-valid twitter value is about to update.
+	 */
+	protected function get_non_valid_twitter( $twitter_site, $twitter_setting ) {
+		if ( $this->options_helper->validate_twitter_id( $twitter_site ) ) {
+			return [];
+		}
+
+		return [ $twitter_setting ];
 	}
 }
